[Kotlin] DSL
A domain-specific language (DSL) is a computer language specialized to a particular application domain.
There are a wide variety of DSLs, ranging from widely used languages for common domains, such as HTML for web pages, down to languages used by only one or a few pieces of software, such as MUSH soft code.